requirements.txt
flask==2.3.3
tensorflow==2.13.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
gunicorn==21.2.0
Werkzeug==2.3.7
lime==0.2.0.1
shap==0.43.0
matplotlib==3.7.2
scikit-image==0.21.0
scikit-learn==1.3.0
